requirements.txt
pillow>=8.1
flask>=1